requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
accelerate==0.28.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
deepspeed==0.16.5
docker-pycreds==0.4.0
evaluate==0.4.3
facenet-pytorch==2.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.29.3
idna==3.10
intel-openmp==2021.4.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lightning-utilities==0.14.3
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
pip==23.2.1
platformdirs==4.3.7
proglog==0.1.11
protobuf==5.29.3
psutil==7.0.0
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tbb==2021.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.2.2
torchmetrics==1.7.0
torchvision==0.17.2
tqdm==4.67.1
transformers==4.38.2
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.9
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2